pyproject.toml
[tool.poetry]
name = "bathytools"
version = "0.1.0"
description = ""
authors = ["Giorgio Bolzon <gbolzon@ogs.it>", "Stefano Piani <spiani@ogs.it>"]
readme = "README.md"
packages = [{include = "bathytools", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
bitsea = { git = "https://github.com/inogs/bit.sea.git", branch = "master" }
pyyaml = "^6.0.2"
numpy = "^2.1.1"
matplotlib = "^3.10.0"
xarray = "^2025.1.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
deptry = "^0.21.2"

[tool.poetry.scripts]
bathytools = "bathytools.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.deptry]
exclude = ["tests/", "examples/"]

[tool.deptry.package_module_name_map]
netcdf4 = "netCDF4"
